Added search to navbar
